fix(hvf): warn when the per-VM supervisor binary is stale
`cargo run` rebuilds only `mvmctl`, never the separate `mvm-hvf-supervisor`
bin that hosts the guest's vsock agent bridge. A stale supervisor boots
cleanly but runs old agent-bridge code, so the guest agent silently never
becomes reachable ("guest agent did not become reachable within 30s") with
no hint at the failing layer — unlike the vz path, where a schema-incompatible
supervisor crashes early and already gets a stale-supervisor hint.

Emit a proactive `ui::warn` on the hvf spawn path (both `HvfBackend::start`
and the `HvfDriver` boot) when the resolved supervisor predates the running
`mvmctl`, naming the exact rebuild command. Lift the vz mtime check into a
shared `supervisor_stale` module so both backends share one implementation.
